Remove mixed precision, added topk pooling, use adamw

+++ train_imagenet.py
@@ -186,8 +186,7 @@ if __name__ == "__main__":
         max_epochs=dict_args["epoch"],
         log_every_n_steps=10,
         gradient_clip_val=dict_args['clip_grad_norm'],
-        devices=-1,
-        precision="16-mixed"
+        devices=-1
     ) 
 
     # Trainer: train model
